+++ edison-jobs/src/test/java/de/otto/edison/jobs/service/JobMetaServiceTest.java
@@ -3,7 +3,7 @@ package de.otto.edison.jobs.service;
 import de.otto.edison.jobs.domain.DisabledJob;
 import de.otto.edison.jobs.domain.RunningJob;
 import de.otto.edison.jobs.repository.JobBlockedException;
-import de.otto.edison.jobs.repository.JobStateRepository;
+import de.otto.edison.jobs.repository.JobMetaRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -23,14 +23,14 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
-public class JobLockServiceTest {
+public class JobMetaServiceTest {
 
     @Mock
-    private JobStateRepository jobStateRepository;
+    private JobMetaRepository jobMetaRepository;
     @Mock
     private JobMutexGroups jobMutexGroups;
 
-    JobLockService jobLockService;
+    JobMetaService jobMetaService;
 
     @Before
     @SuppressWarnings("unchecked")
@@ -38,40 +38,40 @@ public class JobLockServiceTest {
         initMocks(this);
         when(jobMutexGroups.mutexJobTypesFor(anyString()))
                 .thenReturn(emptySet());
-        jobLockService = new JobLockService(jobStateRepository, jobMutexGroups);
+        jobMetaService = new JobMetaService(jobMetaRepository, jobMutexGroups);
     }
 
     @Test
     public void shouldAquireRunLock() throws Exception {
-        when(jobStateRepository.findAllJobTypes()).thenReturn(emptySet());
-        when(jobStateRepository.createValue("myJobType", "_e_running", "jobId")).thenReturn(true);
+        when(jobMetaRepository.findAllJobTypes()).thenReturn(emptySet());
+        when(jobMetaRepository.createValue("myJobType", "_e_running", "jobId")).thenReturn(true);
 
-        jobLockService.aquireRunLock("jobId", "myJobType");
+        jobMetaService.aquireRunLock("jobId", "myJobType");
 
-        verify(jobStateRepository).createValue("myJobType","_e_running", "jobId");
+        verify(jobMetaRepository).createValue("myJobType","_e_running", "jobId");
     }
 
     @Test
     public void shouldReleaseRunLock() {
-        jobLockService.releaseRunLock("someType");
-        verify(jobStateRepository).setValue("someType", "_e_running", null);
+        jobMetaService.releaseRunLock("someType");
+        verify(jobMetaRepository).setValue("someType", "_e_running", null);
     }
 
     @Test(expected = JobBlockedException.class)
     public void shouldNotAquireLockIfAlreadyRunning() throws Exception {
         // given
-        when(jobStateRepository.findAllJobTypes()).thenReturn(emptySet());
+        when(jobMetaRepository.findAllJobTypes()).thenReturn(emptySet());
 
-        when(jobStateRepository.createValue("myJobType", "_e_running", "someId")).thenReturn(false);
+        when(jobMetaRepository.createValue("myJobType", "_e_running", "someId")).thenReturn(false);
 
         // when
         try {
-            jobLockService.aquireRunLock("jobId", "myJobType");
+            jobMetaService.aquireRunLock("jobId", "myJobType");
         }
 
         // then
         catch (final JobBlockedException e) {
-            verify(jobStateRepository, never()).setValue(anyString(), anyString(), anyString());
+            verify(jobMetaRepository, never()).setValue(anyString(), anyString(), anyString());
             throw e;
         }
     }
@@ -79,29 +79,29 @@ public class JobLockServiceTest {
     @Test(expected = JobBlockedException.class)
     public void shouldNotStartJobIfBlockedByAnotherJob() throws Exception {
         // given
-        when(jobStateRepository.findAllJobTypes()).thenReturn(new HashSet<>(asList("job1", "job2")));
-        when(jobStateRepository.getValue("job1", "_e_running")).thenReturn("42");
+        when(jobMetaRepository.findAllJobTypes()).thenReturn(new HashSet<>(asList("job1", "job2")));
+        when(jobMetaRepository.getValue("job1", "_e_running")).thenReturn("42");
         when(jobMutexGroups.mutexJobTypesFor("job2")).thenReturn(new HashSet<>(asList("job1", "job2")));
 
         // when
         try {
-            jobLockService.aquireRunLock("first", "job2");
+            jobMetaService.aquireRunLock("first", "job2");
         }
 
         // then
         catch (final JobBlockedException e) {
-            verify(jobStateRepository, never()).setValue(anyString(), anyString(), anyString());
+            verify(jobMetaRepository, never()).setValue(anyString(), anyString(), anyString());
             throw e;
         }
     }
 
     @Test
     public void shouldReturnRunningJobsDocument() {
-        when(jobStateRepository.findAllJobTypes()).thenReturn(new HashSet<>(asList("someType", "someOtherType")));
-        when(jobStateRepository.getValue("someType", "_e_running")).thenReturn("someId");
-        when(jobStateRepository.getValue("someOtherType", "_e_running")).thenReturn("someOtherId");
+        when(jobMetaRepository.findAllJobTypes()).thenReturn(new HashSet<>(asList("someType", "someOtherType")));
+        when(jobMetaRepository.getValue("someType", "_e_running")).thenReturn("someId");
+        when(jobMetaRepository.getValue("someOtherType", "_e_running")).thenReturn("someOtherId");
 
-        assertThat(jobLockService.runningJobs(), containsInAnyOrder(
+        assertThat(jobMetaService.runningJobs(), containsInAnyOrder(
                 new RunningJob("someId", "someType"),
                 new RunningJob("someOtherId", "someOtherType")
         ));
@@ -110,12 +110,12 @@ public class JobLockServiceTest {
     @Test(expected = JobBlockedException.class)
     public void shouldNotStartADisabledJob() {
         // given
-        when(jobStateRepository.findAllJobTypes()).thenReturn(singleton("jobType"));
-        when(jobStateRepository.getValue("jobType", "_e_disabled")).thenReturn("true");
+        when(jobMetaRepository.findAllJobTypes()).thenReturn(singleton("jobType"));
+        when(jobMetaRepository.getValue("jobType", "_e_disabled")).thenReturn("true");
 
         // when
         try {
-            jobLockService.aquireRunLock("someId", "jobType");
+            jobMetaService.aquireRunLock("someId", "jobType");
         }
 
         // then
@@ -128,13 +128,13 @@ public class JobLockServiceTest {
     @Test
     public void shouldReturnDisabledJobTypes() {
         // given
-        when(jobStateRepository.findAllJobTypes()).thenReturn(new HashSet<String>(asList("jobType", "otherJobType", "enabledJobType")));
-        when(jobStateRepository.getValue("jobType", "_e_disabled")).thenReturn("some comment");
-        when(jobStateRepository.getValue("otherJobType", "_e_disabled")).thenReturn("");
-        when(jobStateRepository.getValue("enabledJobType", "_e_disabled")).thenReturn(null);
+        when(jobMetaRepository.findAllJobTypes()).thenReturn(new HashSet<String>(asList("jobType", "otherJobType", "enabledJobType")));
+        when(jobMetaRepository.getValue("jobType", "_e_disabled")).thenReturn("some comment");
+        when(jobMetaRepository.getValue("otherJobType", "_e_disabled")).thenReturn("");
+        when(jobMetaRepository.getValue("enabledJobType", "_e_disabled")).thenReturn(null);
 
         // when
-        final Set<DisabledJob> disabledJobTypes = jobLockService.disabledJobTypes();
+        final Set<DisabledJob> disabledJobTypes = jobMetaService.disabledJobTypes();
 
         // then
         assertThat(disabledJobTypes, containsInAnyOrder(
@@ -144,20 +144,20 @@ public class JobLockServiceTest {
 
     @Test
     public void shouldDisableJobType() throws Exception {
-        jobLockService.disableJobType(new DisabledJob("jobType", null));
-        verify(jobStateRepository).setValue("jobType", "_e_disabled", "");
+        jobMetaService.disableJobType(new DisabledJob("jobType", null));
+        verify(jobMetaRepository).setValue("jobType", "_e_disabled", "");
     }
 
     @Test
     public void shouldDisableJobTypeWithComment() throws Exception {
-        jobLockService.disableJobType(new DisabledJob("jobType", "some comment"));
-        verify(jobStateRepository).setValue("jobType", "_e_disabled", "some comment");
+        jobMetaService.disableJobType(new DisabledJob("jobType", "some comment"));
+        verify(jobMetaRepository).setValue("jobType", "_e_disabled", "some comment");
     }
 
     @Test
     public void shouldEnableJobType() throws Exception {
-        jobLockService.enableJobType("jobType");
-        verify(jobStateRepository).setValue("jobType", "_e_disabled", null);
+        jobMetaService.enableJobType("jobType");
+        verify(jobMetaRepository).setValue("jobType", "_e_disabled", null);
     }
 
 }