+++ examples/example-jobs/src/main/java/de/otto/edison/example/jobs/ExampleMetaJob.java
@@ -1,9 +1,9 @@
 package de.otto.edison.example.jobs;
 
 import de.otto.edison.jobs.definition.JobDefinition;
-import de.otto.edison.jobs.domain.StatefulJobRunnable;
+import de.otto.edison.jobs.domain.MetaJobRunnable;
 import de.otto.edison.jobs.eventbus.JobEventPublisher;
-import de.otto.edison.jobs.repository.JobStateRepository;
+import de.otto.edison.jobs.service.JobMetaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -12,19 +12,21 @@ import java.util.Optional;
 import static de.otto.edison.jobs.definition.DefaultJobDefinition.manuallyTriggerableJobDefinition;
 
 @Component
-public class StatefulJob extends StatefulJobRunnable {
+public class ExampleMetaJob extends MetaJobRunnable {
+
+    public static final String JOB_TYPE = "ExampleMetaJob";
 
     @Autowired
-    public StatefulJob(final JobStateRepository jobStateRepository) {
-        super(jobStateRepository);
+    public ExampleMetaJob(final JobMetaService jobMetaService) {
+        super(JOB_TYPE, jobMetaService);
     }
 
     @Override
     public JobDefinition getJobDefinition() {
         return manuallyTriggerableJobDefinition(
-                "ExtraValueJob",
-                "ExtraValueJob",
-                "A Job that stores extra values in JobState",
+                JOB_TYPE,
+                "Some stateful Job",
+                "A Job that stores some meta data",
                 0,
                 Optional.empty());
     }
@@ -32,11 +34,11 @@ public class StatefulJob extends StatefulJobRunnable {
     @Override
     public void execute(JobEventPublisher jobEventPublisher) {
 
-        int lastEntry = meta().getAsInt("lastEntry", 0);
+        int lastEntry = jobMeta().getAsInt("lastEntry", 0);
 
         for (int i = lastEntry+1; i <= lastEntry + 10; i++) {
             jobEventPublisher.info("Processing Item " + i);
-            meta().set("lastEntry", i);
+            jobMeta().set("lastEntry", i);
         }
     }
 }